feat(tabs): apply theme
