+++ src/index.ts
@@ -9,7 +9,7 @@ export * from './components/Select/Select';
 export * from './components/Card';
 export * from './components/Radio';
 export * from './components/Skeleton';
-export * from './components/Tabs/Tabs';
+export * from './components/Tabs';
 export * from './components/Table/Table';
 export * from './components/Alert';
 export * from './components/ImageInput';
